Changed plugin path and endpoint hook to be compatible with WP-API (2.0 alpha)
Changed plugin path from "json-rest-api/plugin.php" to "WP-API-develop/plugin.php", so when it checks if the WP-API plugin is activated, it's looking in the right spot. This will need to be updated when 2.0 is released.

Changed "wp_json_server_before_serve" to "wp_json_init" to match the recent naming change in WP-API 2.0 alpha.